Add commands to create and delete remediations (#553)
* Add remediation parameter to login command and bump backplane-api pkg

* add remediation parameter to login

* Add command remediation delete

This can be used to cleanup RBAC created for a remediation given by the
extra parameter in the login flow.

* Adds commands to create and delete remediations

* Fix unit tests

* Refactoring into remediation functions which don't require the full cobra command stack

* Adding tests to remediation

* Fix unit test when run inside container

We are creating kubeconfigs on the filesystem. To allow this to work in
container we can create a temporary kubeconfig.

* Add test for untested function for extracting fields from JWT

These function were moved to the utils package but are in prod for a
long time. I started adding test without changing the behavior for now.

* Move remediation functionality to pkg/remediation

* remove unused return value and fix parameter names

---------

Co-authored-by: Leszek Jakubowski <2552733+Makdaam@users.noreply.github.com>