Make KUBEPS1 warning optional (#692)
* Make KUBEPS1 warning optional
Adds an optional configuration option `disable-kube-ps1-warning` to be able to
conditionally display the warning about setting up the `kube-ps1` function in your terminal.

The motivation for this is I have a custom terminal prompt that includes `~/.kube/config` context
information, and desire for less warning output when I login to clusters.

I tested this behavior with default values (nil), and bool/string, which are handled by
`viper` for us natively.

* Update docs for PS1